requirements.txt
censys==2.1.2
dnspython==2.2.0
requests==2.27.1

